refactor: move input component into chat page. style: improve responsiveness, restyle loading indications

+++ lib/firebaseUtils.ts
@@ -1,10 +1,16 @@
 import {
   addDoc,
   collection,
+  CollectionReference,
   deleteDoc,
   doc,
+  DocumentData,
   serverTimestamp,
   updateDoc,
+  limit,
+  orderBy,
+  query,
+  getDocs,
 } from "firebase/firestore";
 import { db } from "../firebase";
 
@@ -19,8 +25,14 @@ interface NameChatProps {
   name: string;
 }
 
+interface DeleteCollectionProps {
+  collectionRef: CollectionReference<DocumentData>;
+  batchSize: number;
+}
+
 export const createChat = async (email: string | null | undefined) => {
   if (!email) return;
+
   const doc = await addDoc(collection(db, "users", email, "chats"), {
     userId: email,
     createdAt: serverTimestamp(),
@@ -29,11 +41,30 @@ export const createChat = async (email: string | null | undefined) => {
 };
 
 export const deleteChat = async ({ email, id }: DeleteChatProps) => {
-  const docRef = doc(db, "users", email!, "chats", id);
+  if (!email) return;
+
+  const docRef = doc(db, "users", email, "chats", id);
   await deleteDoc(docRef);
 };
 
 export const nameChat = async ({ email, name, id }: NameChatProps) => {
-  const docRef = doc(db, "users", email!, "chats", id);
+  if (!email) return;
+
+  const docRef = doc(db, "users", email, "chats", id);
   if (name) await updateDoc(docRef, { name: name });
+};
+
+export const deleteAllChats = async (email: string | null | undefined) => {
+  if (!email) return;
+
+  const collectionRef = query(collection(db, "users", email, "chats"));
+
+  const chatsToDelete = await getDocs(collectionRef);
+
+  chatsToDelete &&
+    chatsToDelete.forEach((chat) => {
+      const id = chat.id;
+      const docRef = doc(db, "users", email, "chats", id);
+      deleteDoc(docRef);
+    });
 };
